Bring example-forum into the Lesswrong2 repo
This is the start of a refactoring spread over multiple commits, with
the goal of eliminating Vulcan-Starter/example-forum as a distinct
package used by Lesswrong2. The first step (this commit) is to bring
example-forum into the Lesswrong2 repository, so that changes to
Lesswrong2 and to example-forum can be combined into the same commits.
Then we'll have a series of commits to migrate functionality from
example-forum into Lesswrong2 or to remove duplicate functionality from
example-forum. This will eventually result in an example-forum package
which is completely empty, at which point we'll delete it.