Add defensive programming around UsersEmailVerification component
Add a higher-order component `withErrorBoundary`, which wraps other
components in an `<ErrorBoundary>` component, then applies that HoC to
the `UsersEmailVerification` component. Also adds a bit more defensive
programming to Users.emailAddressIsVerified.

This is probably a partial fix for #994, which is related to accounts
with OAuth not having email addresses stored in the right place.
However, I haven't yet succeeded in getting OAuth to work in a context
suitable for testing (account names need to not-collide and Github
immediately marks my test accounts as spam), so it's untested, and
there's more that needs to be done with email-verification and OAuth.