Rewrite FormComponentCheckbox to work nested
When we replaced FormComponentCheckbox from vulcan-ui-bootstrap with our
own implementation, we were sort of guessing about what its API was
supposed to be. We guessed incorrectly, and got a checkbox that works in
forms at the root level, but doesn't work when nested.

The problem is that `props.name` does *not* identify a field on the
document that can be checked and edited, except in the special case
where we're at the root level. Also, it isn't correct to keep a value in
`state`, we should use the one passed as a prop to the form component.

Rewrite FormComponentCheckbox to use the real API. Also be consistent
about what its name is, and call out in a comment that it's a
replacement. Also put //TODOs on other form components that look
suspicious.